refactor: 상태 확인 api router를 / 에서 /status로 변경
