+++ backend/src/app.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { StatusController } from './status/status.controller';
+import { StatusService } from './status/status.service';
 
 @Module({
   imports: [],
-  controllers: [AppController],
-  providers: [AppService],
+  controllers: [AppController, StatusController],
+  providers: [AppService, StatusService],
 })
 export class AppModule {}
