+++ backend/src/app.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class AppService {
-  getMonitorStatus(): string {
-    return JSON.stringify({ statusCode: 200 });
+  getIndex(): string {
+    return 'Hello World!';
   }
 }
