+++ backend/src/status/status.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { StatusService } from './status.service';
+
+@Controller('status')
+export class StatusController {
+  constructor(private readonly statusService: StatusService) {}
+
+  @Get()
+  getMonitorStatus(): string {
+    return this.statusService.getMonitorStatus();
+  }
+}
