+++ backend/src/status/status.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class StatusService {
+  getMonitorStatus(): string {
+    return JSON.stringify({ statusCode: 200 });
+  }
+}
